Move next_sequence_number from State to Service. Tests pass.
