+++ Cargo.toml
@@ -12,7 +12,7 @@ version = "0.8.0"
 [dependencies]
 cbor = "~0.3.16"
 crossbeam = "~0.1.6"
-get_if_addrs = "~0.2.1"
+get_if_addrs = "~0.3.0"
 igd = "~0.4.0"
 ip = "~1.1.0"
 itertools = "~0.4.3"
